Weather App
The Weather Forecast Viewer is a web app that provides accurate weather information for any location. It integrates with the AccuWeather API to display current weather conditions and dynamically change weather icons. Additionally, it shows the current time of the searched location using the Google Time Zone API, enhancing user experience with localized context. With a user-friendly interface and real-time updates, it's a convenient tool for staying informed about weather conditions worldwide.